pulsar-prefect2.0/pyproject.toml
[tool.poetry]
name = "pulsar-prefect2.0"
version = "0.1.0"
description = ""
authors = ["George Coyne <gabcoyne@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
prefect = "2.0a13"
pulsar-client = "2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
